tests/spatnav-distance-function-grid-002.html

<!doctype html>
<meta charset="utf-8">
<title>Spatnav sanity check</title>
<link rel="author" title="Jihye Hong" href="jh.hong@lge.com">
<meta name=flags content="">
<link rel=help href="https://wicg.github.io/spatial-navigation/#high-level-api">

<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<!--load polyfill. To be removed before submitting to WPT-->
<script src="../polyfill/spatnav-heuristic.js"></script>

<style>
  :focus {
    outline: 4px #3582ff solid;
  }

  .container {
    margin: 10px;
    position: relative;
    border: 1px solid black;
    width: 500px;
    height: 300px;
  }

  .box {
    position: absolute;
    box-sizing: border-box;
    width: 50px;
    height: 50px;
    border: 1px black solid;
    background: #6ac;
  }

  #initial_focus {
    background: #F08080;
  }
  
</style>

<body>
  <div class="container">
    <div tabindex="0" class="box" style="left: 220px; top: 40px;"></div>
    <div tabindex="0" class="box" style="left: 280px; top: 40px;"></div>
    <div tabindex="0" class="box" style="left: 340px; top: 40px;"></div>
    <div tabindex="0" class="box" style="left: 320px; top: 175px;"></div>
    <div tabindex="0" class="box" id="right_focus" style="left: 340px; top: 120px;"></div>
    <div tabindex="0" class="box" id="initial_focus" style="width: 200px; left: 50px; top: 120px;"></div>
  </div>  
</body>

<script>
test(function() {
  document.getElementById("initial_focus").focus();
  window.navigate("right");
  assert_equals(document.activeElement.id, "right_focus");
}, "Move focus right from the staring point")
</script>
</html>
